solve merge and wip conflict in #33

+++ protocol/http/message/player.go
@@ -16,7 +16,7 @@ type LoginResponse struct {
 }
 
 type LoginRequest struct {
-	PlayerUID uint   `json:"player_uid"`
+	PlayerUID uint64 `json:"player_uid"`
 	Password  string `json:"password"`
 }
